+++ Vali-Flow.Core/Builder/ValiFlow.cs
@@ -642,10 +642,6 @@ public class ValiFlow<T> : BaseExpression<ValiFlow<T>, T>,
     public ValiFlow<T> AllMatch<TValue>(Expression<Func<T, IEnumerable<TValue>>> selector, ValiFlow<TValue> filter)
         => _collectionExpression.AllMatch(selector, filter);
 
-    [Obsolete("Use Count instead. CountEquals will be removed in a future version.")]
-    public ValiFlow<T> CountEquals<TValue>(Expression<Func<T, IEnumerable<TValue?>>> selector, int count)
-        => _collectionExpression.CountEquals<TValue>(selector, count);
-
     public ValiFlow<T> FutureDate(Expression<Func<T, DateTime>> selector)
         => _dateTimeExpression.FutureDate(selector);
 
@@ -661,14 +657,6 @@ public class ValiFlow<T> : BaseExpression<ValiFlow<T>, T>,
 
     public ValiFlow<T> ExactDate(Expression<Func<T, DateTime>> selector, DateTime date)
         => _dateTimeExpression.ExactDate(selector, date);
-
-    [Obsolete("Use IsBefore instead. BeforeDate will be removed in a future version.")]
-    public ValiFlow<T> BeforeDate(Expression<Func<T, DateTime>> selector, DateTime date)
-        => _dateTimeExpression.BeforeDate(selector, date);
-
-    [Obsolete("Use IsAfter instead. AfterDate will be removed in a future version.")]
-    public ValiFlow<T> AfterDate(Expression<Func<T, DateTime>> selector, DateTime date)
-        => _dateTimeExpression.AfterDate(selector, date);
 
     public ValiFlow<T> IsToday(Expression<Func<T, DateTime>> selector)
         => _dateTimeExpression.IsToday(selector);
@@ -753,14 +741,6 @@ public class ValiFlow<T> : BaseExpression<ValiFlow<T>, T>,
     public ValiFlow<T> ExactDate(Expression<Func<T, DateTimeOffset>> selector, DateTimeOffset date)
         => _dateTimeOffsetExpression.ExactDate(selector, date);
 
-    [Obsolete("Use IsBefore instead. BeforeDate will be removed in a future version.")]
-    public ValiFlow<T> BeforeDate(Expression<Func<T, DateTimeOffset>> selector, DateTimeOffset date)
-        => _dateTimeOffsetExpression.BeforeDate(selector, date);
-
-    [Obsolete("Use IsAfter instead. AfterDate will be removed in a future version.")]
-    public ValiFlow<T> AfterDate(Expression<Func<T, DateTimeOffset>> selector, DateTimeOffset date)
-        => _dateTimeOffsetExpression.AfterDate(selector, date);
-
     public ValiFlow<T> SameMonthAs(Expression<Func<T, DateTimeOffset>> selector, DateTimeOffset date)
         => _dateTimeOffsetExpression.SameMonthAs(selector, date);
 
@@ -826,14 +806,6 @@ public class ValiFlow<T> : BaseExpression<ValiFlow<T>, T>,
 
     public ValiFlow<T> ExactDate(Expression<Func<T, DateOnly>> selector, DateOnly date)
         => _dateOnlyExpression.ExactDate(selector, date);
-
-    [Obsolete("Use IsBefore instead. BeforeDate will be removed in a future version.")]
-    public ValiFlow<T> BeforeDate(Expression<Func<T, DateOnly>> selector, DateOnly date)
-        => _dateOnlyExpression.BeforeDate(selector, date);
-
-    [Obsolete("Use IsAfter instead. AfterDate will be removed in a future version.")]
-    public ValiFlow<T> AfterDate(Expression<Func<T, DateOnly>> selector, DateOnly date)
-        => _dateOnlyExpression.AfterDate(selector, date);
 
     public ValiFlow<T> IsYesterday(Expression<Func<T, DateOnly>> selector)
         => _dateOnlyExpression.IsYesterday(selector);